+++ R/regression_functions.R
@@ -11,19 +11,24 @@ globalVariables(c(
   "rmse",
   "residual",
   "r_squared",
-  "adj_r_squared"
+  "adj_r_squared",
+  "conf_low",
+  "conf_high"
 ))
 
 #' Get regression table
 #'
-#' Wrapper for tidy \code{lm()} regression function output
+#' Output regression table for an \code{lm()} regression in "tidy" format. This function
+#' is a wrapper function for \code{broom::tidy()} and includes confidence
+#' intervals in the output table by default.
 #'
-#' @param model a model object (Currently only \code{lm} is supported)
+#' @param model an \code{lm()} model object
 #' @param digits number of digits precision in output table
 #' @param print If TRUE, return in print format suitable for R Markdown
-#' @param ... other arguments passed to lm()
 #'
-#' @return A tibble or nicely formatted table
+#' @return A tibble-formatted regression table along with lower and upper end 
+#' points of all confidence intervals for all parameters \code{lower_ci} and 
+#' \code{upper_ci}.
 #' @import dplyr
 #' @importFrom stats lm
 #' @importFrom stats predict
@@ -35,15 +40,20 @@ globalVariables(c(
 #' @importFrom janitor clean_names
 #' @importFrom knitr kable
 #' @export
+#' @seealso \code{\link[broom]{tidy}}, \code{\link{get_regression_points}}, \code{\link{get_regression_summaries}}
 #'
 #' @examples
+#' library(moderndive)
+#' 
+#' # Fit lm() regression:
 #' mpg_model <- lm(mpg ~ cyl, data = mtcars)
+#' 
+#' # Get regression table:
 #' get_regression_table(mpg_model)
 get_regression_table <-
   function(model,
            digits = 3,
-           print = FALSE,
-           ...) {
+           print = FALSE) {
     
     input_checks(model, digits, print)
     
@@ -55,7 +65,11 @@ get_regression_table <-
       mutate_if(is.numeric, round, digits = digits) %>%
       mutate(term = ifelse(term == "(Intercept)", "intercept", term)) %>%
       as_tibble() %>%
-      clean_names()
+      clean_names() %>% 
+      rename(
+        lower_ci = conf_low,
+        upper_ci = conf_high
+      )
     
     if(print) {
       regression_table <- regression_table %>%
@@ -68,12 +82,18 @@ get_regression_table <-
 
 #' Get regression points
 #'
-#' Wrapper for tidy \code{lm()} regression function fit output
+#' Output information on each point/observation used in an \code{lm()} regression in 
+#' "tidy" format. This function is a wrapper function for \code{broom::augment()} 
+#' and renames the variables to have more intuitive names.
 #'
 #' @inheritParams get_regression_table
-#' @param newdata a new data frame to fit the regression to/make predictions on.
+#' @param newdata A new data frame of points/observations to apply \code{model} to 
+#' obtain new fitted values and/or predicted values y-hat. Note the format of 
+#' \code{newdata} must match the format of the original \code{data} used to fit
+#' \code{model}.
 #'
-#' @return A tibble or nicely formatted table
+#' @return A tibble-formatted regression table of outcome variable, 
+#' all explanatory/predictor variables, the fitted/predicted value, and residual.
 #' @import dplyr
 #' @import rlang
 #' @importFrom stats formula 
@@ -87,28 +107,31 @@ get_regression_table <-
 #' @importFrom knitr kable
 #' @importFrom rlang sym
 #' @export
+#' @seealso \code{\link[broom]{augment}}, \code{\link{get_regression_table}}, \code{\link{get_regression_summaries}}
 #'
 #' @examples
+#' library(moderndive)
 #' library(dplyr)
 #' library(tibble)
 #' 
-#' # Get information on all points in regression
+#' # Fit lm() regression:
 #' mpg_model <- lm(mpg ~ cyl, data = mtcars)
+#' 
+#' # Get information on all points in regression:
 #' get_regression_points(mpg_model)
 #' 
-#' # Convert rownames to column
-#' data("mtcars")
+#' # Create training and test set based on mtcars: 
 #' mtcars <- mtcars %>% 
 #'   rownames_to_column(var = "model")
-#'
-#' # Create training and test set.
 #' training_set <- mtcars %>% 
 #'   sample_frac(0.5)
 #' test_set <- mtcars %>% 
 #'   anti_join(training_set, by = "model")
 #' 
-#' # Fit model to training set, predict on test set
+#' # Fit model to training set:
 #' mpg_model_train <- lm(mpg ~ cyl, data = training_set)
+#' 
+#' # Make predictions on test set:
 #' get_regression_points(mpg_model_train, newdata = test_set)
 get_regression_points <-
   function(model,
@@ -168,11 +191,12 @@ get_regression_points <-
 
 #' Get regression summary values
 #'
-#' Wrapper for \code{lm()} regression function fit summary data
+#' Output scalar summary statistics for an \code{lm()} regression in "tidy" 
+#' format. This function is a wrapper function for \code{broom::glance()}.
 #'
 #' @inheritParams get_regression_table
 #'
-#' @return data frame with summaries
+#' @return A single-row tibble with regression summaries. Ex: \code{r_squared} and \code{mse}.
 #' @import dplyr
 #' @importFrom stats formula
 #' @importFrom magrittr "%>%"
@@ -184,15 +208,20 @@ get_regression_points <-
 #' @importFrom janitor clean_names
 #' @importFrom knitr kable
 #' @export
+#' @seealso \code{\link[broom]{glance}}, \code{\link{get_regression_table}}, \code{\link{get_regression_points}}
 #'
 #' @examples
+#' library(moderndive)
+#' 
+#' # Fit lm() regression:
 #' mpg_model <- lm(mpg ~ cyl, data = mtcars)
+#' 
+#' # Get regression summaries:
 #' get_regression_summaries(mpg_model)
 get_regression_summaries <-
   function(model,
            digits = 3,
-           print = FALSE,
-           ...) {
+           print = FALSE) {
     
     input_checks(model, digits, print)
     
@@ -226,8 +255,7 @@ get_regression_summaries <-
 
 input_checks <- function(model,
                          digits = 3,
-                         print = FALSE,
-                         ...){
+                         print = FALSE){
   # Since the `"glm"` class also contains the `"lm"` class
   if(length(class(model)) != 1 | !("lm" %in% class(model)) ){
     stop(paste("Only simple and multiple linear regression",