+++ NEWS.md
@@ -1,9 +1,18 @@
 # moderndive 0.1.1.9000
 
-* Import `infer::rep_sample_n()` instead of our own defined version, as this function is now in `infer` on CRAN
-* Added `evals` and `house_prices` datasets
+Updated package for:
+
+- Use in DataCamp's [Modeling with Data in the Tidyverse](https://www.datacamp.com/courses/modeling-with-data-in-the-tidyverse), in particular added `evals` and `house_prices` datasets and updated `get_regression_table()` and `get_regression_points()` functions.
+- v0.4.0 of [ModernDive](https://moderndive.github.io/moderndive_book/) textbook
+
+Details:
+
+* Import `infer::rep_sample_n()` instead of our own defined version, as this function is [now included in `infer`](https://github.com/andrewpbray/infer/pull/82)
+* Added `evals`, `house_prices`, `tactile_prop_red` datasets
 * Added mean squared error and root mean squared error to output of `get_regression_summaries()`
-* Added `newdata` argument to `get_regression_points()`. When outcome variable is in `newdata`, output the outcome variable's observed value, fitted value, and residual (See Issue 17).
+* Added `newdata` argument to `get_regression_points()`. When:
+    - Original outcome variable is included in `newdata`, output it as well as `residual` (See Issue 17).
+    - Otherwise omit `residual`
 * Removed `tidyverse` from Depends, Imports, or Suggests
 
 # moderndive 0.1.1